Create impls for Printf and VerifStatement
