+++ core/src/main/scala-2/chisel3/Printf.scala
@@ -13,32 +13,10 @@ import scala.language.experimental.macros
   *
   * See apply methods for use
   */
-object printf {
+object printf extends PrintfImpl {
 
   /** Named class for [[printf]]s. */
   final class Printf private[chisel3] (val pable: Printable) extends VerificationStatement
-
-  /** Helper for packing escape characters */
-  private[chisel3] def format(formatIn: String): String = {
-    require(formatIn.forall(c => c.toInt > 0 && c.toInt < 128), "format strings must comprise non-null ASCII values")
-    def escaped(x: Char) = {
-      require(x.toInt >= 0, s"char ${x} to Int ${x.toInt} must be >= 0")
-      if (x == '"' || x == '\\') {
-        s"\\${x}"
-      } else if (x == '\n') {
-        "\\n"
-      } else if (x == '\t') {
-        "\\t"
-      } else {
-        require(
-          x.toInt >= 32,
-          s"char ${x} to Int ${x.toInt} must be >= 32"
-        ) // TODO \xNN once FIRRTL issue #59 is resolved
-        x
-      }
-    }
-    formatIn.map(escaped).mkString("")
-  }
 
   /** Prints a message in simulation
     *
