+++ core/src/main/scala-2/chisel3/VerificationStatement.scala
@@ -5,15 +5,9 @@ package chisel3
 import chisel3.experimental.{BaseModule, SourceInfo}
 import chisel3.internal._
 import chisel3.internal.Builder.pushCommand
-import chisel3.internal.firrtl.ir._
 import chisel3.layer.block
 
 import scala.language.experimental.macros
-
-/** Base class for all verification statements: Assert, Assume, Cover, Stop and Printf. */
-abstract class VerificationStatement extends NamedComponent {
-  _parent.foreach(_.addId(this))
-}
 
 /** Scaladoc information for internal verification statement macros
   * that are used in objects assert, assume and cover.
@@ -29,7 +23,7 @@ abstract class VerificationStatement extends NamedComponent {
   */
 trait VerifPrintMacrosDoc
 
-object assert extends VerifPrintMacrosDoc {
+object assert extends assertImpl with VerifPrintMacrosDoc {
 
   /** Named class for assertions. */
   final class Assert private[chisel3] () extends VerificationStatement
@@ -79,15 +73,9 @@ object assert extends VerifPrintMacrosDoc {
 
   def apply(cond: Bool)(implicit sourceInfo: SourceInfo): Assert =
     macro VerifStmtMacrosCompat.assert._applyMacroWithNoMessage
-
-  /** An elaboration-time assertion. Calls the built-in Scala assert function. */
-  def apply(cond: Boolean, message: => String): Unit = Predef.assert(cond, message)
-
-  /** An elaboration-time assertion. Calls the built-in Scala assert function. */
-  def apply(cond: Boolean): Unit = Predef.assert(cond, "")
 }
 
-object assume extends VerifPrintMacrosDoc {
+object assume extends assumeImpl with VerifPrintMacrosDoc {
 
   /** Named class for assumptions. */
   final class Assume private[chisel3] () extends VerificationStatement
@@ -141,20 +129,12 @@ object assume extends VerifPrintMacrosDoc {
 
   def apply(cond: Bool)(implicit sourceInfo: SourceInfo): Assume =
     macro VerifStmtMacrosCompat.assume._applyMacroWithNoMessage
-
-  /** An elaboration-time assumption. Calls the built-in Scala assume function. */
-  def apply(cond: Boolean, message: => String): Unit = Predef.assume(cond, message)
-
-  /** An elaboration-time assumption. Calls the built-in Scala assume function. */
-  def apply(cond: Boolean): Unit = Predef.assume(cond, "")
 }
 
 object cover extends VerifPrintMacrosDoc {
 
   /** Named class for cover statements. */
   final class Cover private[chisel3] () extends VerificationStatement
-
-  type SourceLineInfo = (String, Int)
 
   /** Declares a condition to be covered.
     * At ever clock event, a counter is incremented iff the condition is active
@@ -176,6 +156,9 @@ object cover extends VerifPrintMacrosDoc {
 
 object stop {
 
+  /** Named class for [[stop]]s. */
+  final class Stop private[chisel3] () extends VerificationStatement
+
   /** Terminate execution, indicating success and printing a message.
     *
     * @param message a message describing why simulation was stopped
@@ -191,9 +174,6 @@ object stop {
   /** Terminate execution, indicating success.
     */
   def apply()(implicit sourceInfo: SourceInfo): Stop = buildStopCommand(None)
-
-  /** Named class for [[stop]]s. */
-  final class Stop private[chisel3] () extends VerificationStatement
 
   private def buildStopCommand(message: Option[Printable])(implicit sourceInfo: SourceInfo): Stop = {
     val stopId = new Stop()
