+++ core/src/main/scala/chisel3/VerificationStatementImpl.scala
@@ -1,0 +1,33 @@
+// SPDX-License-Identifier: Apache-2.0
+
+package chisel3
+
+import chisel3.experimental.SourceInfo
+import chisel3.internal.NamedComponent
+
+/** Base class for all verification statements: Assert, Assume, Cover, Stop and Printf. */
+abstract class VerificationStatement extends NamedComponent {
+  _parent.foreach(_.addId(this))
+}
+
+private[chisel3] trait assertImpl {
+
+  /** An elaboration-time assertion. Calls the built-in Scala assert function. */
+  def apply(cond: Boolean, message: => String): Unit = Predef.assert(cond, message)
+
+  /** An elaboration-time assertion. Calls the built-in Scala assert function. */
+  def apply(cond: Boolean): Unit = Predef.assert(cond, "")
+}
+
+private[chisel3] trait assumeImpl {
+
+  /** An elaboration-time assumption. Calls the built-in Scala assume function. */
+  def apply(cond: Boolean, message: => String): Unit = Predef.assume(cond, message)
+
+  /** An elaboration-time assumption. Calls the built-in Scala assume function. */
+  def apply(cond: Boolean): Unit = Predef.assume(cond, "")
+}
+
+private[chisel3] trait coverImpl {
+  type SourceLineInfo = (String, Int)
+}
